feat/rename reducer
